% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipebook.R
\name{create_recipebook}
\alias{create_recipebook}
\title{Create a recipebook - a tibble storing recipes}
\usage{
create_recipebook(names = character(), ingredients = list())
}
\arguments{
\item{ingredients}{}
}
\value{
A \code{tibble::tibble()} containing recipe details.
}
\description{
Create a recipebook - a tibble storing recipes
}
\examples{

test_ingred <- Ingredients(
    names = c(
        "chicken",
        "salt",
        "pepper"
    ),
    amounts = c(1, 2, 3),
    units = c(NA_character_, "g", "g")
)

create_recipebook(c("Roast Chicken"), list(test_ingred))
}
