% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopping_list.R
\name{create_shopping_list}
\alias{create_shopping_list}
\title{Create a shopping list}
\usage{
create_shopping_list(recipebook)
}
\arguments{
\item{recipebook}{\code{dplyr::tibble()} containing recipe information.}
}
\value{
\code{tibble()} containing the shopping list.
}
\description{
Create a shopping list
}
\examples{

meal_plan <- create_meal_plan(recipebook_example)
shopping_list <- create_shopping_list(meal_plan)

shopping_list
}
