% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ingredients-class.R, R/Ingredients-methods.R
\docType{class}
\name{Ingredients-class}
\alias{Ingredients-class}
\alias{Ingredients}
\alias{names,Ingredients-method}
\alias{amounts}
\alias{units}
\title{An S4 class representing a set of ingredients.}
\usage{
Ingredients(names, amounts = 1, units = rep(NA_character_, length(names)))

\S4method{names}{Ingredients}(x)

amounts(x)

units(x)
}
\arguments{
\item{names}{\code{character()} name of each ingredient.}

\item{amounts}{\code{numeric()} amount of each ingredient.}

\item{units}{\code{character()} units for the amount of each ingredient.}

\item{x}{of class \code{Ingredients-class}}
}
\description{
Contains information regarding the name, the quantity and the
units for each ingredient.
}
\section{Functions}{
\itemize{
\item \code{names,Ingredients-method}: obtain names of ingredients

\item \code{amounts}: obtain amounts of ingredients

\item \code{units}: obtain units of amounts for each ingredient
}}

\section{Slots}{

\describe{
\item{\code{names}}{name of each ingredient.}

\item{\code{amounts}}{amount of each ingredient.}

\item{\code{units}}{units for the amount of each ingredient.}
}}

