% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ingredients.R
\name{read_ingredients}
\alias{read_ingredients}
\title{Read in a list of ingredients}
\usage{
read_ingredients(ingredients, delim = NULL, method = "auto")
}
\arguments{
\item{ingredients}{\code{character()} containing ingredient info separated by
\code{delim}. Each element must denote the ingredients coming from each recipe.}

\item{delim}{\code{character()} that separates each ingredient.}

\item{method}{\code{character()} used to determine the method of parsing
ingredients, placeholder for future implementations (e.g. with data.frame
input).}
}
\value{
\code{list()} of elements of \code{Ingredients-class}.
}
\description{
Read in a list of ingredients
}
\examples{

read_ingredients(
    c(
        "butter;6 eggs;1 bag coriander;1 punnet mixed baby tomatoes",
        "1 tsp garam masala;1 bag baby spinach;1 lemon;1 chilli"
    ),
    delim = ";"
)
}
