% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\alias{RecipeBook}
\title{An S4 class representing a set of recipes}
\usage{
RecipeBook(names, ingredients)
}
\description{
Contains recipe info.
}
