% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recipebook_example}
\alias{recipebook_example}
\title{An example recipe book}
\format{
A \code{tibble::tibble()} with 26 rows and 2 columns:
\describe{
\item{names}{\code{character()} recipe name.}
\item{ingredients}{
\code{character()} name, quantity and type of ingredient, stored as a \code{list()}
of \code{Ingredients-class} objects.
}
}
}
\source{
generated using data-raw/recipebook_example.R
}
\usage{
recipebook_example
}
\description{
A data set that has been created using \code{autorecipes::create_recipebook()}
containing 26 example recipes. The recipes originate from recipe boxes that
I've ordered from \href{https://www.riverford.co.uk}{Riverford}. Riverford are a
wonderful UK company producing sustainable, delicious, locally-sourced food -
check them out!
}
\keyword{datasets}
