% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_recipes.R
\name{set_recipes}
\alias{set_recipes}
\alias{add_recipes}
\alias{rm_recipes}
\title{Add or remove recipes}
\usage{
add_recipes(recipebook, names, ingredients)

rm_recipes(recipebook, indexes = NULL, method = c("manual", "index"))
}
\arguments{
\item{recipebook}{object of \code{RecipeBook-class}.}

\item{names}{\code{character()} containing the name of of recipes.}

\item{ingredients}{\code{character()} or \code{list()} of \code{Ingredient-class} objects.
If a \code{character()}, will be read in using \code{read_ingredients()}.}

\item{indexes}{\code{numeric()} the indexes of the recipes to change, required if
method is set to 'index'.}

\item{method}{\code{character()} one of 'manual' or 'index'. If 'manual', will
allow users to interactively pick the recipes to add/remove as favourites.
If 'index' will modify the recipes corresponding to \code{indexes}.}
}
\description{
Add or remove recipes
}
