% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{weekdays}
\alias{weekdays}
\title{Obtain the weekdays of interest}
\usage{
weekdays(which_days = NULL)
}
\arguments{
\item{which_days}{\code{numeric()} or \code{character()}, default \code{NULL}. If numeric,
must be an index between 1:7. If character, must be one of  \code{c("Mon", "Tues", "Wed", "Thurs", "Fri", "Sat", "Sun")}. If NULL, will return all
week days.}
}
\value{
\code{character()} with selected weekdays
}
\description{
Obtain the weekdays of interest
}
\examples{

weekdays()
weekdays(1:3)
weekdays(c("Mon", "Tues", "Wed"))
}
