% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal_plan.R
\name{create_meal_plan}
\alias{create_meal_plan}
\title{Create a meal plan}
\usage{
create_meal_plan(recipebook, method = c("auto", "random"))
}
\arguments{
\item{recipebook}{\code{tibble::tibble()} containing the recipe information,
possibly created using \code{create_recipebook()}.}

\item{method}{\code{character()} the method to use when creating a meal plan.}
}
\value{
\code{tibble::tibble()} containing a meal plan.
}
\description{
Create a meal plan
}
\examples{
meal_plan <- create_meal_plan(recipebook_example)

meal_plan
}
