% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_meal_plan.R
\name{create_meal_plan}
\alias{create_meal_plan}
\title{Create a customisable meal plan}
\usage{
create_meal_plan(
  recipebook,
  which_days = weekdays(),
  which_meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE,
  set_last_eaten = TRUE
)
}
\arguments{
\item{recipebook}{a \code{RecipeBook-class} object.}

\item{which_days}{\code{character()} which days of the week to plan for, in the
format "Mon"}

\item{which_meals}{\code{character()} which meals to plan for, one of "Lunch" or
"Dinner".}

\item{method}{\code{character()} the method to use when creating a meal plan, one
of "auto", or "random".}

\item{fav_only}{\code{logical()} whether to only pick from favourite recipes.}

\item{set_last_eaten}{\code{logical()} whether to record the that selected recipes
were used for a meal plan today. This info is taken into account when
optimising the meal planning algorithm if \code{method} is set to "auto".}
}
\value{
a \code{RecipeBook-class} object containing the created \code{meal_plan}.
}
\description{
\code{create_meal_plan()} expects a \code{RecipeBook-class} object as input. The
available options for customizing the meal plan include; the days of the week
and the meals ("Lunch" or "Dinner") to plan for, the method to use and
whether to only use favourite recipes.
}
\examples{
recipebook <- create_meal_plan(recipebook_example)

meal_plan(recipebook)
}
