% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal_plan.R
\name{create_meal_plan}
\alias{create_meal_plan}
\title{Create a meal plan}
\usage{
create_meal_plan(
  recipebook,
  days = c("Mon", "Tues", "Wed", "Thurs", "Fri", "Sat", "Sun"),
  meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE
)
}
\arguments{
\item{recipebook}{\code{RecipeBook} object containing the recipe information,
possibly created using \code{create_recipebook()}.}

\item{days}{\code{character()} which days of the week to plan for.}

\item{meals}{\code{character()} which meals to plan for, out of "Lunch" and
"Dinner".}

\item{method}{\code{character()} the method to use when creating a meal plan.}

\item{fav_only}{\code{logical()} whether to create a meal plan only using
favourite recipes.}
}
\value{
\code{tibble::tibble()} containing a meal plan.
}
\description{
Create a meal plan
}
\examples{
meal_plan <- create_meal_plan(recipebook_example)

meal_plan
}
