% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal_plan.R
\name{create_meal_plan}
\alias{create_meal_plan}
\title{Create a meal plan}
\usage{
create_meal_plan(
  recipebook,
  days = c("Mon", "Tues", "Wed", "Thurs", "Fri", "Sat", "Sun"),
  meals = c("Lunch", "Dinner"),
  method = c("auto", "random"),
  fav_only = FALSE,
  set_last_eaten = TRUE
)
}
\arguments{
\item{recipebook}{a \code{RecipeBook-class} object.}

\item{days}{\code{character()} which days of the week to plan for.}

\item{meals}{\code{character()} which meals to plan for, out of "Lunch" and
"Dinner".}

\item{method}{\code{character()} the method to use when creating a meal plan.}

\item{fav_only}{\code{logical()} whether to create a meal plan only using
favourite recipes.}

\item{set_last_eaten}{\code{logical()} whether to save the recipes
were eaten today. Used to optimise the meal planning algorithm when the
\code{method} is set to "auto".}
}
\value{
a \code{RecipeBook-class} object containing the created \code{meal_plan}.
}
\description{
Create a meal plan
}
\examples{
recipebook <- create_meal_plan(RecipeBook_example)

meal_plan(recipebook)
}
