% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meal_plan.R
\name{create_meal_plan}
\alias{create_meal_plan}
\title{Create a meal plan}
\usage{
create_meal_plan(recipebook, type = c("auto", "random"))
}
\arguments{
\item{recipebook}{\code{tibble::tibble()} containing the recipes from which meal
plan will be created.}

\item{type}{\code{character()} the type of meal plan to create.}
}
\value{
A meal plan
}
\description{
Create a meal plan
}
