% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_shopping_list.R
\name{create_shopping_list}
\alias{create_shopping_list}
\title{Create a shopping list}
\usage{
create_shopping_list(
  recipebook,
  filter_method = c("none", "manual", "minimal")
)
}
\arguments{
\item{recipebook}{a \code{RecipeBook-class} object.}

\item{filter_method}{\code{character()} the method to used to filter out store
cupboard ingredients from your shopping list.}
}
\value{
\code{tibble::tibble()} containing a shopping list.
}
\description{
Expects a \code{RecipeBook-class} object with non-empty meal plan slot as input. A
shopping list consist of a summarised table with all of the ingredients that
are required to cook the recipes in the current meal plan.
}
