% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shopping_list.R
\name{create_shopping_list}
\alias{create_shopping_list}
\title{Create a shopping list}
\usage{
create_shopping_list(recipebook)
}
\arguments{
\item{recipebook}{a \code{RecipeBook-class} object.}
}
\value{
\code{tibble::tibble()} containing a shopping list.
}
\description{
Create a shopping list
}
