% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ingredients-class.R, R/Ingredients-methods.R
\docType{class}
\name{Ingredients-class}
\alias{Ingredients-class}
\alias{Ingredients}
\alias{show,Ingredients-method}
\alias{names,Ingredients-method}
\alias{amounts}
\alias{units}
\alias{as.data.frame,Ingredients-method}
\title{S4 class containing names, amounts and units for ingredients}
\usage{
Ingredients(names, amounts = 1, units = rep(NA_character_, length(names)))

\S4method{show}{Ingredients}(object)

\S4method{names}{Ingredients}(x)

amounts(x)

units(x)

\S4method{as.data.frame}{Ingredients}(x)
}
\arguments{
\item{names}{\code{character()} name of each ingredient.}

\item{amounts}{\code{numeric()} amount of each ingredient.}

\item{units}{\code{character()} units for the amount of each ingredient.}

\item{object}{an \code{Ingredients-class} object.}

\item{x}{an \code{Ingredients-class} object.}
}
\description{
Contains information regarding the name, the quantity and the
units for each ingredient.
}
\section{Slots}{

\describe{
\item{\code{names}}{name of each ingredient.}

\item{\code{amounts}}{amount of each ingredient.}

\item{\code{units}}{units for the amount of each ingredient.}
}}

\section{Constructor}{


\code{Ingredients(names, amounts, units)} creates an object of
\code{Ingredients-class}.
}

\section{Displaying}{


\code{show(ingredients)} prints each ingredient.
}

\section{Getters}{


\code{names(ingredients)} obtain the names of each ingredient.



\code{amounts(ingredients)} obtain the amounts of each ingredient.



\code{units(ingredients)} obtain the units of each ingredient.



\code{as.data.frame(ingredients)} converts ingredients to a \code{data.frame}.
}

