% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ingredients-class.R, R/Ingredients-methods.R
\docType{class}
\name{Ingredients-class}
\alias{Ingredients-class}
\alias{show,Ingredients-method}
\alias{names,Ingredients-method}
\alias{amounts,Ingredients-method}
\alias{units,Ingredients-method}
\title{An S4 class representing a set of ingredients.}
\usage{
\S4method{show}{Ingredients}(object)

\S4method{names}{Ingredients}(x)

\S4method{amounts}{Ingredients}(x)

\S4method{units}{Ingredients}(x)
}
\description{
An S4 class representing a set of ingredients.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print set of ingredients

\item \code{names}: obtain the names of ingredients

\item \code{amounts}: obtain the amount of ingredients

\item \code{units}: obtain the units of the amount
}}

\section{Slots}{

\describe{
\item{\code{names}}{name of each ingredient.}

\item{\code{amounts}}{amount of each ingredient.}

\item{\code{units}}{units for the amount of each ingredient.}
}}

