% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R, R/RecipeBook-methods.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\alias{RecipeBook}
\alias{show,RecipeBook-method}
\alias{names,RecipeBook-method}
\alias{recipes}
\alias{meal_plan}
\alias{favourites}
\alias{favourites<-}
\title{RecipeBook containing recipes and ingredients}
\usage{
\S4method{show}{RecipeBook}(object)

\S4method{names}{RecipeBook}(x)

recipes(x)

meal_plan(x)

favourites(x)

favourites(x) <- value
}
\arguments{
\item{object}{RecipeBook object to display.}

\item{x}{a \code{RecipeBook} object.}
}
\description{
A book of recipes from which meal plans can be generated.
}
\section{Displaying}{


\code{show(recipebook)} displays the \code{recipes} slot as a \code{tibble::tibble()}.
}

\section{Getters}{


\code{names(recipebook)} obtains the names of the recipes.



\code{recipes(recipebook)} obtains the \code{recipes} as a \code{tibble::tibble()}.



\code{meal_plan(recipebook)} obtains the \code{meal_plan} as a \code{tibble::tibble()}.



\code{favourites(recipebook)} obtains the \code{recipes} marked as a favourite.
}

\section{Setters}{


\code{favourites(recipebook) <- fav_indexes} set your favourite \code{recipes}.
}

