% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R, R/RecipeBook-methods.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\alias{RecipeBook}
\alias{show,RecipeBook-method}
\alias{names,RecipeBook-method}
\alias{length,RecipeBook-method}
\alias{recipes}
\alias{meal_plan}
\alias{shopping_list}
\alias{favourites}
\alias{recipes<-}
\title{S4 class containing recipes, their ingredients and meal plans}
\usage{
RecipeBook(names, ingredients)

\S4method{show}{RecipeBook}(object)

\S4method{names}{RecipeBook}(x)

\S4method{length}{RecipeBook}(x)

recipes(object)

meal_plan(object)

shopping_list(object)

favourites(x)

recipes(object) <- value
}
\arguments{
\item{names}{\code{character()} containing the name of of recipes.}

\item{ingredients}{\code{character()} or \code{list()} of \code{Ingredient-class} objects.
If a \code{character()}, will be read in using \code{read_ingredients()}.}

\item{object}{a \code{RecipeBook-class} object.}

\item{x}{a \code{RecipeBook-class} object.}

\item{value}{\code{tibble::tibble()} which will be used to set the \code{recipes} slot.}
}
\description{
An \code{RecipeBook-class} object contains 3 components; 1. a set of recipes
retrievable using \code{recipes()}, 2. a meal plan created using
\code{create_meal_plan()} and 3. a shopping list created using
\code{create_shopping_list()}.
}
\section{Slots}{

\describe{
\item{\code{recipes}}{\code{tibble::tibble()} storing the names, ingredients and
favourites status of recipes.}

\item{\code{meal_plan}}{\code{tibble::tibble()} storing a meal plan, created by
\code{create_meal_plan()}.}

\item{\code{shopping_list}}{\code{tibble::tibble()} storing a shopping list, created by
\code{create_shopping_list()}.}
}}

\section{Constructor}{


\code{RecipeBook(names, ingredients)} creates an object of \code{RecipeBook-class}.
}

\section{Displaying}{


\code{show(recipebook)} prints the \code{recipes} as a \code{tibble::tibble()}.
}

\section{Getters}{


\code{names(recipebook)} obtains the names of the recipes.



\code{length(recipebook)} obtains the number of recipes.



\code{recipes(recipebook)} obtains the \code{recipes} as a \code{tibble::tibble()}.



\code{meal_plan(recipebook)} obtains the \code{meal_plan} as a \code{tibble::tibble()}.



\code{shopping_list(recipebook)} obtains the \code{shopping_list} as a
\code{tibble::tibble()}.



\code{favourites(recipebook)} obtains the \code{recipes} marked as favourites.
}

\section{Setters}{


\code{recipes(object) <- value} can be used to set the \code{recipes}.
}

