% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\title{Title}
\value{

}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{recipes}}{DFrame.}

\item{\code{meal_plan}}{DFrame.}
}}

