% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R, R/RecipeBook-methods.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\alias{RecipeBook}
\alias{names,RecipeBook-method}
\alias{ingredients,RecipeBook-method}
\title{An S4 class representing a set of recipes}
\usage{
RecipeBook(names, ingredients)

\S4method{names}{RecipeBook}(x)

\S4method{ingredients}{RecipeBook}(x)
}
\arguments{
\item{names}{\code{character()} name of each recipe.}

\item{ingredients}{\code{list()} set of ingredients required for each recipe.}
}
\description{
Name, ingredients of recipes.
}
\section{Functions}{
\itemize{
\item \code{names,RecipeBook-method}: obtain names of ingredients

\item \code{ingredients,RecipeBook-method}: obtain ingredients for each recipe
}}

\section{Slots}{

\describe{
\item{\code{names}}{name of each recipe.}

\item{\code{ingredients}}{set of ingredients required for each recipe.}
}}

