% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RecipeBook-class.R, R/RecipeBook-methods.R,
%   R/favourites.R
\docType{class}
\name{RecipeBook-class}
\alias{RecipeBook-class}
\alias{RecipeBook}
\alias{show,RecipeBook-method}
\alias{names,RecipeBook-method}
\alias{length,RecipeBook-method}
\alias{recipes}
\alias{meal_plan}
\alias{meal_plan<-}
\alias{favourites}
\alias{favourites<-}
\alias{favourites<-,RecipeBook,numeric-method}
\alias{favourites<-,RecipeBook,logical-method}
\alias{favourites<-,RecipeBook,character-method}
\title{S4 class containing recipes, their ingredients and meal plans}
\usage{
RecipeBook(names, ingredients)

\S4method{show}{RecipeBook}(object)

\S4method{names}{RecipeBook}(x)

\S4method{length}{RecipeBook}(x)

recipes(object)

meal_plan(object)

meal_plan(object) <- value

favourites(x)

favourites(object, ...) <- value

\S4method{favourites}{RecipeBook,numeric}(object) <- value

\S4method{favourites}{RecipeBook,logical}(object) <- value

\S4method{favourites}{RecipeBook,character}(object, con = stdin()) <- value
}
\arguments{
\item{names}{\code{character()} containing the name of of recipes.}

\item{ingredients}{\code{character()} or \code{list()} of \code{Ingredient-class} objects.
If a \code{character()}, will be read in using \code{read_ingredients()}.}

\item{object}{a \code{RecipeBook-class} object.}

\item{x}{a \code{RecipeBook-class} object.}

\item{value}{either a \code{integer()} specifying the index of the rows to
favourite, an \code{logical()} to set as the "fav" column, or the \code{character()}
"manual" to specify setting the favourites manually.}

\item{...}{additional arguments passed on the \verb{favourite<-} methods.}

\item{con}{a \link[base]{connection} object or a character string.}
}
\description{
A book of recipes from which meal plans and shopping lists can be generated.
}
\section{Slots}{

\describe{
\item{\code{recipes}}{\code{tibble::tibble()} storing the names, ingredients and
favourites status of recipes.}

\item{\code{meal_plan}}{\code{tibble::tibble()} storing a meal plan, created by \code{recipes}
using \code{create_meal_plan()}.}
}}

\section{Constructor}{


\code{RecipeBook(names, ingredients)} creates an object of \code{RecipeBook-class}.
}

\section{Displaying}{


\code{show(recipebook)} prints the \code{recipes} as a \code{tibble::tibble()}.
}

\section{Getters}{


\code{names(recipebook)} obtains the names of the recipes.



\code{length(recipebook)} obtains the number of recipes.



\code{recipes(recipebook)} obtains the \code{recipes} as a \code{tibble::tibble()}.



\code{meal_plan(recipebook)} obtains the \code{meal_plan} as a \code{tibble::tibble()}.



\code{favourites(recipebook)} obtains the \code{recipes} marked as favourites.
}

\section{Setters}{


\code{meal_plan(recipebook) <- meal_plan} stores the meal plan inside a
\code{RecipeBook-class} object.



\code{favourites(recipebook) <- 1} set favourite \code{recipes} by index.



\code{favourites(recipebook) <- TRUE} set favourite \code{recipes} by logical.



\code{favourites(recipebook) <- "manual"} set favourite \code{recipes} manually.
}

