% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{recipes_example}
\alias{recipes_example}
\title{Set of example example recipes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26 rows and 3 columns.
}
\source{
generated using data-raw/recipes_examples.R
}
\usage{
recipes_example
}
\description{
A data set containing 25 example recipes originating from my (recipe-box
inspired) cupboard.
}
\details{
#' @format A \code{tibble} with 53940 rows and 10 variables:
\describe{
\item{title}{\code{character()} recipe name.}
\item{title_tidy}{\code{character()} recipe name in computer-friendly format. }
\item{ingredients}{
\code{character()} quantity and type of ingredient, separated by a ";"
}
}
}
\keyword{datasets}
