% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_favourites.R
\name{set_favourites}
\alias{set_favourites}
\title{Pick your favourite recipes}
\usage{
set_favourites(recipebook, fav_indexes = NULL)
}
\arguments{
\item{recipebook}{\code{tibble::tibble()} containing the recipe information,
possibly created using \code{create_recipebook()}.}

\item{fav_indexes}{\code{integer()} set of indexes that correspond to the recipes
that you want to favourite.}
}
\value{
\code{tibble::tibble()} containing recipebook with selected favourites.
}
\description{
Pick your favourite recipes
}
\examples{
set_favourites(recipebook_example, fav_indexes = c(1, 2, 3))
}
